[arccore,common+accelerator] Use 'Int64' instead of 'size_t' for sizes in 'IMemoryPool'.
This is to be coherent with the other methods about memory.

+++ arcane/src/arcane/utils/tests/TestMemoryPool.cc
@@ -1,6 +1,6 @@
 ﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
 //-----------------------------------------------------------------------------
-// Copyright 2000-2025 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
+// Copyright 2000-2026 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: Apache-2.0
 //-----------------------------------------------------------------------------
@@ -25,8 +25,8 @@ using namespace Arcane::Impl;
 class MyMemoryPoolAllocator
 : public IMemoryPoolAllocator
 {
-  void* allocateMemory(size_t size) override { return std::malloc(size); }
-  void freeMemory(void* address,size_t) override { std::free(address); }
+  void* allocateMemory(Int64 size) override { return std::malloc(size); }
+  void freeMemory(void* address,Int64) override { std::free(address); }
 };
 
 /*---------------------------------------------------------------------------*/
